- Add support for homestead and ropsten ABI files - Ropsten ABI is "pre-8" - Homestead ABI per 20.12.2018 deployment - Refactoring: Move wallet related files into subfolder
