+++ lib/payment.spec.js
@@ -12,7 +12,7 @@ const MonetaryAmount = require('./monetary-amount');
 const {privateToAddress} = require('ethereumjs-util');
 const {hash, prefix0x} = require('./utils');
 
-const Wallet = proxyquire('./wallet', {
+const Wallet = proxyquire('./wallet/wallet', {
     './client-fund-contract': function() {
         return {};
     },
