+++ lib/receipt.spec.js
@@ -20,7 +20,7 @@ const sender = prefix0x(privateToAddress(new Buffer(senderKey, 'hex')).toString(
 const operatorKey = '252c772d9e9d570c726473927c52f23227cc674b6567935287d3f4a76a57d352';
 const operatorAddress = prefix0x(privateToAddress(new Buffer(operatorKey, 'hex')).toString('hex'));
 
-const Wallet = proxyquire('./wallet', {
+const Wallet = proxyquire('./wallet/wallet', {
     './client-fund-contract': function() {
         return {};
     },
