+++ lib/settlement.spec.js
@@ -46,7 +46,7 @@ const stubbedNullSettlement = {
 };
 
 function proxyquireWallet() {
-    return proxyquire('./wallet', {
+    return proxyquire('./wallet/wallet', {
         './client-fund-contract': function() {
             return stubbedClientFundContract;
         },
