+++ lib/wallet/balance-tracker-contract.js
@@ -1,13 +1,15 @@
 'use strict';
 
 const ethers = require('ethers');
-const balanceTrackerDeployment = require('./abis/BalanceTracker');
 
 class BalanceTrackerContract extends ethers.Contract {
     constructor(walletOrProvider) {
         const provider = walletOrProvider.provider || walletOrProvider;
+        const networkName = provider.network.name;
+        const balanceTrackerDeployment = require(`./abis/${networkName}/BalanceTracker`);
         const chainId = provider.network.chainId;
         const balanceTrackerAddress = balanceTrackerDeployment.networks[chainId].address;
+
         super(balanceTrackerAddress, balanceTrackerDeployment.abi, walletOrProvider);
     }
 }
