+++ lib/wallet/balance-tracker-contract.spec.js
@@ -30,12 +30,13 @@ const fakeEthers = {
 
 const BalanceTrackerContract = proxyquire('./balance-tracker-contract', {
     'ethers': fakeEthers,
-    './abis/BalanceTracker': fakeBalanceTrackerDeployment
+    './abis/ropsten/BalanceTracker': fakeBalanceTrackerDeployment
 });
 
 const fakeProvider = {
     network: {
-        chainId: '123456789'
+        chainId: '123456789',
+        name: 'ropsten'
     }
 };
 const fakeWallet = {
