+++ lib/wallet/client-fund-contract.js
@@ -1,13 +1,15 @@
 'use strict';
 
 const ethers = require('ethers');
-const clientFundDeployment = require('./abis/ClientFund');
 
 class ClientFundContract extends ethers.Contract {
     constructor(walletOrProvider) {
         const provider = walletOrProvider.provider || walletOrProvider;
+        const networkName = provider.network.name;
+        const clientFundDeployment = require(`./abis/${networkName}/ClientFund`);
         const chainId = provider.network.chainId;
         const clientFundAddress = clientFundDeployment.networks[chainId].address;
+
         super(clientFundAddress, clientFundDeployment.abi, walletOrProvider);
     }
 }
