+++ lib/wallet/client-fund-contract.spec.js
@@ -30,12 +30,13 @@ const fakeEthers = {
 
 const ClientFundContract = proxyquire('./client-fund-contract', {
     'ethers': fakeEthers,
-    './abis/ClientFund': fakeClientFundDeployment
+    './abis/ropsten/ClientFund': fakeClientFundDeployment
 });
 
 const fakeProvider = {
     network: {
-        chainId: '123456789'
+        chainId: '123456789',
+        name: 'ropsten'
     }
 };
 const fakeWallet = {
