+++ lib/wallet/erc20-contract.js
