+++ lib/wallet/wallet.js
@@ -4,7 +4,7 @@
  * @module nahmii-sdk
  */
 
-const dbg = require('./dbg');
+const dbg = require('../dbg');
 const ethers = require('ethers');
 const ClientFundContract = require('./client-fund-contract');
 const BalanceTrackerContract = require('./balance-tracker-contract');
@@ -167,11 +167,11 @@ class Wallet extends ethers.Signer {
         const tokenContract = new Erc20Contract(tokenInfo.currency, this);
         const amountBN = ethers.utils.parseUnits(amount.toString(), tokenInfo.decimals);
         const clientFund = _clientFund.get(this);
-        
+
         try {
             return await clientFund.receiveTokens('', amountBN, tokenContract.address, 0, 'ERC20', options);
         }
-        catch(e) {
+        catch (e) {
             throw new Error(`Failed to complete token deposit: ${e}`);
         }
     }
@@ -194,7 +194,8 @@ class Wallet extends ethers.Signer {
     }
 
     /**
-     * Unstage an amount of ETH or ERC20 tokens from staged balance back to nahmii available balance.
+     * Unstage an amount of ETH or ERC20 tokens from staged balance back to
+     * nahmii available balance.
      * @param {MonetaryAmount} monetaryAmount - The amount unstage from staged balance.
      * @param [options]
      * @return {Promise} A promise that resolves into transaction hash.
@@ -219,9 +220,9 @@ class Wallet extends ethers.Signer {
     }
 
     /**
-     * Signs message and returns a Promise that resolves to the flat-format signature.
-     * If message is a string, it is converted to UTF-8 bytes, otherwise it is preserved 
-     * as a binary representation of the Arrayish data.
+     * Signs message and returns a Promise that resolves to the flat-format
+     * signature. If message is a string, it is converted to UTF-8 bytes,
+     * otherwise it is preserved as a binary representation of the Arrayish data.
      * @param message
      * @return {Promise<string>}
      */
@@ -230,8 +231,10 @@ class Wallet extends ethers.Signer {
     }
 
     /**
-     * Signs transaction and returns a Promise that resolves to the signed transaction as a hex string.
-     * In general, the sendTransaction method is preferred to sign, as it can automatically populate values asynchronously.
+     * Signs transaction and returns a Promise that resolves to the signed
+     * transaction as a hex string.
+     * In general, the sendTransaction method is preferred to sign, as it can
+     * automatically populate values asynchronously.
      * @param transaction
      * @return {Promise<string>}
      */
@@ -258,8 +261,8 @@ class Wallet extends ethers.Signer {
     }
 
     /**
-     * Sends the transaction to the network and returns 
-     * a Promise that resolves to a Transaction Response. 
+     * Sends the transaction to the network and returns a Promise that resolves
+     * to a Transaction Response.
      * Any properties that are not provided will be populated from the network.
      * See: https://docs.ethers.io/ethers.js/html/api-providers.html#transaction-request
      * @param {object} transaction - An unsigned Ethereum transaction
@@ -270,8 +273,8 @@ class Wallet extends ethers.Signer {
     }
 
     /**
-     * If used with software wallet, returns an object contianing signer related information and logic 
-     * such as the private key, otherwise undefined
+     * If used with software wallet, returns an object containing signer related
+     * information and logic such as the private key, otherwise undefined
      * @returns {ethers.SignerKey|undefined} The private key or undefined
      */
     get signerKey() {
