+++ lib/wallet/wallet.spec.js
@@ -8,7 +8,7 @@ chai.use(sinonChai);
 
 const proxyquire = require('proxyquire').noPreserveCache().noCallThru();
 const ethers = require('ethers');
-const MonetaryAmount = require('./monetary-amount');
+const MonetaryAmount = require('../monetary-amount');
 
 const privateKey = '0x' + '0F'.repeat(32);
 const walletAddress = '0x691A8D05678FC962ff0f2174134379c0051Cb686'; // Derived from privateKey! Not random!
@@ -78,7 +78,7 @@ function proxyquireWallet() {
         address: walletAddress,
         signMessage: async (rawMessage) => {
             let message = rawMessage;
-            if (typeof message === 'string') 
+            if (typeof message === 'string')
                 message = ethers.utils.toUtf8Bytes(message);
             const wallet = new ethers.Wallet(privateKey);
             const signature = wallet.signingKey.signDigest(ethers.utils.hashMessage(message));
@@ -88,9 +88,9 @@ function proxyquireWallet() {
             const tx = await ethers.utils.resolveProperties(unresolvedTx);
             const serializedTx = ethers.utils.serializeTransaction(tx);
             const wallet = new ethers.Wallet(privateKey);
-            const signature = wallet.signingKey.signDigest(ethers.utils.keccak256(serializedTx)); 
+            const signature = wallet.signingKey.signDigest(ethers.utils.keccak256(serializedTx));
             return ethers.utils.serializeTransaction(tx, signature);
-        } 
+        }
     }]
 ].forEach(([description, signer]) => {
     describe(`Wallet with ${description}`, () => {
@@ -306,7 +306,7 @@ function proxyquireWallet() {
                     stubbedClientFundContract.receiveTokens
                         .withArgs('', ethers.utils.parseUnits('2.71', t.decimals), t.currency, 0, 'ERC20', {gasLimit: 1})
                         .resolves({hash: fakeTxHash});
-                    
+
                     const {hash} = await wallet.completeTokenDeposit('2.71', t.symbol, {gasLimit: 1});
                     expect(hash).to.eql(fakeTxHash);
                 });
